+++ src/apps/hasabe/components/General/CardButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./CardButton";
