+++ src/apps/hasabe/components/General/UtilityBar/UtilityBar.css
@@ -1,0 +1,15 @@
+.UtilityBar {
+    position: sticky;
+    top: 0;
+    padding: var(--app-padding);
+    margin: calc(-1 * var(--app-padding));
+    z-index: 100;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    background-image: linear-gradient(var(--color-background-trans), transparent);
+}
+
+.UtilityBar > *:not(:last-child) {
+    margin-right: 8px;
+}
