+++ src/apps/hasabe/components/General/UtilityBar/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./UtilityBar";
