+++ src/apps/hasabe/components/General/index.ts
@@ -1,0 +1,1 @@
+export { default as CardButton } from "./CardButton";
