+++ src/apps/hasabe/components/Tags/index.ts
@@ -1,0 +1,1 @@
+export { default as TagList } from "./TagList";
