--- src/apps/hasabe/components/TaskListActions.css
@@ -1,12 +1,0 @@
-.TaskListActions {
-    position: sticky;
-    top: 16px;
-    z-index: 100;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.TaskListActions > *:not(:last-child) {
-    margin-right: 8px;
-}
